requirements.txt
asgiref==3.3.4
astroid==2.5.6
boto3==1.17.66
botocore==1.20.66
colorama==0.4.4
Django==3.2
django-environ==0.4.5
django-storages==1.11.1
djangorestframework==3.12.4
environ==1.0
isort==5.8.0
jmespath==0.10.0
lazy-object-proxy==1.6.0
logilab-common==1.8.1
mccabe==0.6.1
mypy-extensions==0.4.3
mysql-connector-python==8.0.24
mysqlclient==2.0.3
Pillow==8.2.0
protobuf==3.15.8
psycopg2==2.8.6
pylint==2.8.2
pylint-django==2.4.4
pylint-plugin-utils==0.6
python-dateutil==2.8.1
pytz==2021.1
s3transfer==0.4.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.4
wrapt==1.12.1
